TASK#320: Implements the export to PDF feature for a review.
